wasserstoff/AiTask/requirements.txt
chromadb==0.4.24
sentence-transformers==2.2.2
langchain==0.1.17
langchain-chroma==0.1.0
langchain-community==0.0.37
langchain-core==0.1.52
langchain-google-genai==1.0.3
python-dotenv==1.0.1
streamlit==1.34.0
google-generativeai==0.5.3
